initial stepper implementation
